requirements.txt
djangorestframework==3.6.2
Django==1.11
Pygments==2.1.3
Markdown==2.6.3
django-multiselectfield

# heroku
# run these commands to setup heroku properly
# heroku create --stack cedar --buildpack git://github.com/heroku/heroku-buildpack-python.git
# heroku config:add BUILDPACK_URL=git@github.com:heroku/heroku-buildpack-python.git#purge
# heroku config:set HEROKU=1
psycopg2==2.6.1
dj-database-url==0.3.0
